Update README.md
DESCRIPTION:
This repository contains a Solidity smart contract named MyToken, implementing a basic token system with minting and burning functionalities. The contract includes public variables for token details such as name, abbreviation, and total supply. It uses a mapping to manage the balances of different addresses. The mint function allows increasing the total supply and the balance of a specified address, while the burn function decreases the total supply and balance of a specified address, ensuring sufficient tokens are available to burn.

AUTHOR
Swati Kumari
swaditya.sk54@gmail.com

LICENSE
The project is licensed under the MIT License.